pyproject.toml
[project]
name = "mcp-server-code-assist"
version = "0.2.0"
description = "MCP Code Assist Server"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.0.0",
    "gitpython>=3.1.40",
    "pydantic>=2.0.0",
    "click>=8.1.7",
    "mcp>=1.2.0",
    "xmlschema>=3.4.3"
]

[project.scripts]
mcp-server-code-assist = "mcp_server_code_assist.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.2",
    "pre-commit>=4.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "strict"
testpaths = ["tests"]

[tool.ruff]
line-length = 200
indent-width = 4
fix = true
preview = true

lint.select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "W",  # pycodestyle warnings
    "C90",  # Complexity
    "C",  # flake8-comprehensions
    "ISC",  # flake8-implicit-str-concat
    "T10",  # flake8-debugger
    "A",  # flake8-builtins
    "UP",  # pyupgrade
]

lint.ignore = [
    "C416",  # Unnecessary list comprehension - rewrite as a generator expression
    "C408"  # Unnecessary `dict` call - rewrite as a literal
]

lint.fixable = ["ALL"]
lint.unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
combine-as-imports = true
